Experimental code to test basic login functionality
Plugin Updates

Add code to test login and API access

Fix karakitsario

First steps towards the API Wrapper for Reddit that fox is going to use. This class aims to abstract(kind of) requests to the Reddit API, for now.

Refactor so this uses foxApiWrapper.lib.RedditRequest to talk to Reddit. Pretty much a playground until we reach a point where the wrapper is ready to be plugged to a real UI.